+++ src/app/routes/private/routes/eventos.private.js
@@ -1,0 +1,7 @@
+const express = require('express')
+const route = express.Router()
+const EventoController = require('./../../../controllers/eventos')
+
+route.post('/criar', EventoController.create)
+
+module.exports =  route